chore: apply lint fixes and document pnpm reinstall

+++ database/migrations/2025_11_18_000001_create_node_snapshots_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
 
         Schema::create('node_snapshots', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('node_id')->index();
+            $table->unsignedInteger('node_id')->index();
             $table->timestamp('recorded_at')->index();
 
             // Basic utilization
